refactor(fusion): the deferred drops are a file, not a nested module
They have their own state, their own doc and their own tests, which is
what a file is. `tensor.rs` becomes `tensor/base.rs` beside it, the
layout every other module in this crate already uses.
