+++ crates/burn-fusion/src/tensor/mod.rs
@@ -1,0 +1,5 @@
+mod base;
+
+pub(crate) mod deferred;
+
+pub use base::*;
